+++ pkg/controller/informers_test.go
@@ -99,12 +99,13 @@ var _ = Describe("Informers Tests", func() {
 					VirtualServerAddress: "1.2.3.5",
 					Partition:            "dev",
 				})
-			mockCtlr.resources.ltmConfig[mockCtlr.Partition] = &PartitionConfig{}
+			zero := 0
+			mockCtlr.resources.ltmConfig[mockCtlr.Partition] = &PartitionConfig{ResourceMap: make(ResourceMap), Priority: &zero}
 			mockCtlr.enqueueUpdatedVirtualServer(vs, newVS)
 			key, quit = mockCtlr.resourceQueue.Get()
 			Expect(key).ToNot(BeNil(), "Enqueue Updated VS Failed")
 			Expect(quit).To(BeFalse(), "Enqueue Updated VS  Failed")
-			Expect(mockCtlr.resources.ltmConfig[mockCtlr.Partition].Priority).To(BeEquivalentTo(1), "Priority Not Updated")
+			Expect(*mockCtlr.resources.ltmConfig[mockCtlr.Partition].Priority).To(BeEquivalentTo(1), "Priority Not Updated")
 			delete(mockCtlr.resources.ltmConfig, mockCtlr.Partition)
 			key, quit = mockCtlr.resourceQueue.Get()
 			Expect(key).ToNot(BeNil(), "Enqueue Updated VS Failed")
@@ -163,6 +164,33 @@ var _ = Describe("Informers Tests", func() {
 			mockCtlr.enqueueVirtualServer(vs)
 			Expect(mockCtlr.processResources()).To(Equal(true))
 
+			// Verify VS status update event is not queued for processing
+			updatedStatusVS := test.NewVirtualServer(
+				"SampleVS",
+				namespace,
+				cisapiv1.VirtualServerSpec{
+					Host:                 "test.com",
+					VirtualServerAddress: "5.6.7.8",
+					SNAT:                 "none",
+				})
+			updatedStatusVS.Status.StatusOk = "OK"
+			mockCtlr.enqueueUpdatedVirtualServer(updatedVS2, updatedStatusVS)
+			Expect(mockCtlr.resourceQueue.Len()).To(Equal(0), "VS status update should be skipped")
+
+			// Verify VS Label update event is queued for processing
+			updatedLabelVS := test.NewVirtualServer(
+				"SampleVS",
+				namespace,
+				cisapiv1.VirtualServerSpec{
+					Host:                 "test.com",
+					VirtualServerAddress: "5.6.7.8",
+					SNAT:                 "none",
+				})
+			labels := make(map[string]string)
+			labels["f5cr"] = "false"
+			updatedLabelVS.Labels = labels
+			mockCtlr.enqueueUpdatedVirtualServer(updatedStatusVS, updatedLabelVS)
+			Expect(mockCtlr.resourceQueue.Len()).To(Equal(1), "VS label update should not be skipped")
 		})
 
 		It("TLS Profile", func() {
@@ -222,9 +250,26 @@ var _ = Describe("Informers Tests", func() {
 			Expect(mockCtlr.processResources()).To(Equal(true))
 			tsWithPartition := newTS.DeepCopy()
 			tsWithPartition.Spec.Partition = "dev"
-			mockCtlr.resources.ltmConfig[mockCtlr.Partition] = &PartitionConfig{}
+			zero := 0
+			mockCtlr.resources.ltmConfig[mockCtlr.Partition] = &PartitionConfig{ResourceMap: make(ResourceMap), Priority: &zero}
 			mockCtlr.enqueueUpdatedTransportServer(newTS, tsWithPartition)
-			Expect(mockCtlr.resources.ltmConfig[mockCtlr.Partition].Priority).To(BeEquivalentTo(1), "Priority Not Updated")
+			Expect(*mockCtlr.resources.ltmConfig[mockCtlr.Partition].Priority).To(BeEquivalentTo(1), "Priority Not Updated")
+
+			// Verify TS status update event is not queued for processing
+			queueLen := mockCtlr.resourceQueue.Len()
+			updatedStatusTS := tsWithPartition.DeepCopy()
+			updatedStatusTS.Status.StatusOk = "Ok"
+			mockCtlr.enqueueUpdatedTransportServer(tsWithPartition, updatedStatusTS)
+			Expect(mockCtlr.resourceQueue.Len()).To(Equal(queueLen), "TS status update should be skipped")
+
+			// Verify TS Label update event is queued for processing
+			updatedLabelTS := updatedStatusTS.DeepCopy()
+			labels := make(map[string]string)
+			labels["f5cr"] = "false"
+			updatedLabelTS.Labels = labels
+			mockCtlr.enqueueUpdatedTransportServer(updatedStatusTS, updatedLabelTS)
+			Expect(mockCtlr.resourceQueue.Len()).To(Equal(queueLen+1), "TS label update should not be skipped")
+
 		})
 
 		It("IngressLink", func() {
@@ -279,9 +324,10 @@ var _ = Describe("Informers Tests", func() {
 
 			ilWithPartition := newIL.DeepCopy()
 			ilWithPartition.Spec.Partition = "dev"
-			mockCtlr.resources.ltmConfig[mockCtlr.Partition] = &PartitionConfig{}
+			zero := 0
+			mockCtlr.resources.ltmConfig[mockCtlr.Partition] = &PartitionConfig{ResourceMap: make(ResourceMap), Priority: &zero}
 			mockCtlr.enqueueUpdatedIngressLink(newIL, ilWithPartition)
-			Expect(mockCtlr.resources.ltmConfig[mockCtlr.Partition].Priority).To(BeEquivalentTo(1), "Priority Not Updated")
+			Expect(*mockCtlr.resources.ltmConfig[mockCtlr.Partition].Priority).To(BeEquivalentTo(1), "Priority Not Updated")
 
 		})
 
@@ -365,6 +411,13 @@ var _ = Describe("Informers Tests", func() {
 		})
 
 		It("Service", func() {
+			// setting teem data
+			mockCtlr.TeemData = &teem.TeemsData{
+				ResourceType: teem.ResourceTypes{
+					IPAMSvcLB:   make(map[string]int),
+					IngressLink: make(map[string]int),
+				},
+			}
 			svc := test.NewService(
 				"SampleSVC",
 				"1",