+++ pkg/controller/nativeResourceWorker_test.go
@@ -501,7 +501,8 @@ extendedRouteSpec:
 			//Process ENDS with non-matching domain
 			mockCtlr.addEDNS(newEDNS)
 			mockCtlr.processExternalDNS(newEDNS, false)
-			gtmConfig := mockCtlr.resources.gtmConfig[DEFAULT_PARTITION].WideIPs
+			gtmConfig := mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs
+			DEFAULT_GTM_PARTITION = DEFAULT_GTM_PARTITION + "_gtm"
 			Expect(len(gtmConfig)).To(Equal(1))
 			Expect(len(gtmConfig["test.com"].Pools)).To(Equal(1))
 			// No pool member should be present
@@ -510,14 +511,14 @@ extendedRouteSpec:
 			//delete EDNS
 			mockCtlr.deleteEDNS(newEDNS)
 			mockCtlr.processExternalDNS(newEDNS, true)
-			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_PARTITION].WideIPs
+			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs
 			Expect(len(gtmConfig)).To(Equal(0))
 
 			// Modify EDNS with matching domain and create again
 			mockCtlr.addEDNS(newEDNS)
 			newEDNS.Spec.DomainName = "pytest-foo-1.com"
 			mockCtlr.processExternalDNS(newEDNS, false)
-			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_PARTITION].WideIPs
+			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs
 			Expect(len(gtmConfig)).To(Equal(1))
 			Expect(len(gtmConfig["pytest-foo-1.com"].Pools)).To(Equal(1))
 			// Pool member should be present
@@ -527,7 +528,7 @@ extendedRouteSpec:
 			mockCtlr.deleteRoute(route1)
 			mockCtlr.deleteHostPathMapEntry(route1)
 			mockCtlr.processRoutes(namespace1, false)
-			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_PARTITION].WideIPs
+			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs
 			Expect(len(gtmConfig)).To(Equal(1))
 			Expect(len(gtmConfig["pytest-foo-1.com"].Pools)).To(Equal(1))
 			// No pool member should be present
@@ -582,7 +583,7 @@ extendedRouteSpec:
 			//Test with 2nd route with bigIpPartition
 			mockCtlr.addEDNS(barEDNS)
 			mockCtlr.processExternalDNS(barEDNS, false)
-			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_PARTITION].WideIPs
+			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs
 			Expect(len(gtmConfig)).To(Equal(2))
 			Expect(len(gtmConfig["pytest-bar-1.com"].Pools)).To(Equal(1))
 			Expect(len(gtmConfig["pytest-bar-1.com"].Pools[0].Members)).To(Equal(1))
@@ -590,7 +591,7 @@ extendedRouteSpec:
 
 			mockCtlr.deleteEDNS(barEDNS)
 			mockCtlr.processExternalDNS(barEDNS, true)
-			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_PARTITION].WideIPs
+			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs
 			Expect(len(gtmConfig)).To(Equal(1))
 
 			//Remove route group
@@ -607,7 +608,7 @@ extendedRouteSpec:
 			Expect(err).To(BeNil())
 			Expect(isProcessed).To(BeTrue())
 
-			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_PARTITION].WideIPs
+			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs
 			Expect(len(gtmConfig)).To(Equal(1))
 			Expect(len(gtmConfig["pytest-foo-1.com"].Pools)).To(Equal(1))
 			//No pool members should present
@@ -674,8 +675,23 @@ extendedRouteSpec:
 					DestinationCACertificate:      "",
 				},
 			}
+			spec3 := routeapi.RouteSpec{
+				Host: "baz.com",
+				Path: "/baz",
+				To: routeapi.RouteTargetReference{
+					Kind: "Service",
+					Name: "baz",
+				},
+				TLS: &routeapi.TLSConfig{Termination: "edge",
+					Certificate:                   "",
+					Key:                           "",
+					InsecureEdgeTerminationPolicy: "",
+					DestinationCACertificate:      "",
+				},
+			}
 			route1 := test.NewRoute("route1", "1", routeGroup, spec1, nil)
 			route2 := test.NewRoute("route2", "1", routeGroup, spec2, nil)
+			route3 := test.NewRoute("route3", "1", routeGroup, spec3, nil)
 
 			// Resource Config for unsecured virtual server
 			rsCfg := &ResourceConfig{}
@@ -707,6 +723,12 @@ extendedRouteSpec:
 			Expect(len(rsCfg.Policies[0].Rules)).To(Equal(1))
 			Expect(rsCfg.Policies[0].Rules[0].FullURI).To(Equal("foo.com/foo"))
 
+			// HTTP virtual server, secured route, InsecureEdgeTerminationPolicy = ""
+			Expect(mockCtlr.prepareResourceConfigFromRoute(rsCfg, route3, intstr.IntOrString{IntVal: 80}, ps)).To(BeNil())
+			Expect(rsCfg.Policies).NotTo(BeNil())
+			Expect(len(rsCfg.Policies)).To(Equal(1))
+			Expect(len(rsCfg.Policies[0].Rules)).To(Equal(1))
+
 			// ResourceConfig for secured virtual server
 			rsCfg = &ResourceConfig{}
 			rsCfg.Virtual.Partition = routeGroup
@@ -727,8 +749,14 @@ extendedRouteSpec:
 			Expect(rsCfg.Policies).NotTo(BeNil())
 			Expect(len(rsCfg.Policies)).To(Equal(1))
 			Expect(len(rsCfg.Policies[0].Rules)).To(Equal(2))
-			Expect(rsCfg.Policies[0].Rules[0].FullURI).To(Equal("foo.com/foo"))
-			Expect(rsCfg.Policies[0].Rules[1].FullURI).To(Equal("bar.com/bar"))
+			Expect(mockCtlr.prepareResourceConfigFromRoute(rsCfg, route3, intstr.IntOrString{IntVal: 80}, ps)).To(BeNil())
+			Expect(rsCfg.Policies).NotTo(BeNil())
+			Expect(len(rsCfg.Policies)).To(Equal(1))
+			Expect(len(rsCfg.Policies[0].Rules)).To(Equal(3))
+			// Check Rules are in sorted order
+			Expect(rsCfg.Policies[0].Rules[0].FullURI).To(Equal("bar.com/bar"))
+			Expect(rsCfg.Policies[0].Rules[1].FullURI).To(Equal("baz.com/baz"))
+			Expect(rsCfg.Policies[0].Rules[2].FullURI).To(Equal("foo.com/foo"))
 
 		})