+++ pkg/controller/worker_test.go
@@ -189,7 +189,7 @@ var _ = Describe("Worker Tests", func() {
 	})
 
 	Describe("IPAM", func() {
-		DEFAULT_PARTITION = "Test"
+		DEFAULT_PARTITION = "test"
 		BeforeEach(func() {
 			mockCtlr.Agent = &Agent{
 				PostManager: &PostManager{
@@ -926,6 +926,7 @@ var _ = Describe("Worker Tests", func() {
 		It("Processing External DNS", func() {
 			mockCtlr.resources.Init()
 			DEFAULT_PARTITION = "default"
+			DEFAULT_GTM_PARTITION = "default_gtm"
 			mockCtlr.TeemData = &teem.TeemsData{
 				ResourceType: teem.ResourceTypes{
 					ExternalDNS: make(map[string]int),
@@ -952,26 +953,27 @@ var _ = Describe("Worker Tests", func() {
 					},
 				})
 			mockCtlr.processExternalDNS(newEDNS, false)
-			gtmConfig := mockCtlr.resources.gtmConfig[DEFAULT_PARTITION].WideIPs
+			gtmConfig := mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs
 			Expect(len(gtmConfig)).To(Equal(1))
 			Expect(len(gtmConfig["test.com"].Pools)).To(Equal(1))
 			Expect(len(gtmConfig["test.com"].Pools[0].Members)).To(Equal(0))
 			Expect(gtmConfig["test.com"].Pools[0].Ratio).To(Equal(4))
 
-			mockCtlr.resources.ltmConfig["default"] = &PartitionConfig{make(ResourceMap), 0}
+			zero := 0
+			mockCtlr.resources.ltmConfig["default"] = &PartitionConfig{ResourceMap: make(ResourceMap), Priority: &zero}
 			mockCtlr.resources.ltmConfig["default"].ResourceMap["SampleVS"] = &ResourceConfig{
 				MetaData: metaData{
 					hosts: []string{"test.com"},
 				},
 			}
 			mockCtlr.processExternalDNS(newEDNS, false)
-			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_PARTITION].WideIPs
+			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs
 			Expect(len(gtmConfig)).To(Equal(1))
 			Expect(len(gtmConfig["test.com"].Pools)).To(Equal(1))
 			Expect(len(gtmConfig["test.com"].Pools[0].Members)).To(Equal(1))
 
 			mockCtlr.processExternalDNS(newEDNS, true)
-			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_PARTITION].WideIPs
+			gtmConfig = mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs
 			Expect(len(gtmConfig)).To(Equal(0))
 		})
 
@@ -1966,6 +1968,10 @@ var _ = Describe("Worker Tests", func() {
 				vs31.Spec.Partition = "dev"
 				mockCtlr.updateVirtualServer(&vs3, &vs31)
 				mockCtlr.processResources()
+				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(2), "Invalid Partition count")
+				// Simulating partition priority update to zero by response handler on successfully posting the priority
+				// tenant update
+				mockCtlr.resources.updatePartitionPriority("dev3", 0)
 				mockCtlr.processResources()
 				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(1), "Invalid Partition count")
 				Expect(len(mockCtlr.resources.ltmConfig["dev"].ResourceMap)).To(Equal(2), "Invalid VS count")
@@ -1986,6 +1992,10 @@ var _ = Describe("Worker Tests", func() {
 				vs2.Spec.Host = "zya.com"
 				mockCtlr.updateVirtualServer(vs, &vs2)
 				mockCtlr.processResources()
+				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(2), "Invalid Partition count")
+				// Simulating partition priority update to zero by response handler on successfully posting the priority
+				// tenant update
+				mockCtlr.resources.updatePartitionPriority("dev", 0)
 				mockCtlr.processResources()
 				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(1), "Invalid Partition count")
 				Expect(len(mockCtlr.resources.ltmConfig[mockCtlr.Partition].ResourceMap)).To(Equal(2), "Invalid VS count")
@@ -2259,7 +2269,7 @@ var _ = Describe("Worker Tests", func() {
 				newTS2.Spec.VirtualServerAddress = "10.1.1.3"
 				newTS2.Spec.VirtualServerName = "ts_crd_2"
 				mockCtlr.addTransportServer(&newTS2)
-				// Should not process TS now. Already one TS with same IP is present in different partition
+				// Should process this TS as it has unique IP and vs name
 				mockCtlr.processResources()
 				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(2), "Invalid Partition count")
 				Expect(len(mockCtlr.resources.ltmConfig["dev2"].ResourceMap)).To(Equal(1), "Invalid TS count")
@@ -2282,6 +2292,10 @@ var _ = Describe("Worker Tests", func() {
 				newTS31.Spec.Partition = "dev2"
 				mockCtlr.updateTransportServer(&newTS3, &newTS31)
 				mockCtlr.processResources()
+				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(3), "Invalid Partition count")
+				// Simulating partition priority update to zero by response handler on successfully posting the priority
+				// tenant update
+				mockCtlr.resources.updatePartitionPriority("dev3", 0)
 				mockCtlr.processResources()
 				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(2), "Invalid Partition count")
 				Expect(len(mockCtlr.resources.ltmConfig["dev"].ResourceMap)).To(Equal(1), "Invalid TS count")
@@ -2293,6 +2307,10 @@ var _ = Describe("Worker Tests", func() {
 				newTS32.Spec.Partition = ""
 				mockCtlr.updateTransportServer(&newTS31, &newTS32)
 				mockCtlr.processResources()
+				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(2), "Invalid Partition count")
+				// Simulating partition priority update to zero by response handler on successfully posting the priority
+				// tenant update
+				mockCtlr.resources.updatePartitionPriority("dev2", 0)
 				mockCtlr.processResources()
 				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(3), "Invalid Partition count")
 				Expect(len(mockCtlr.resources.ltmConfig[mockCtlr.Partition].ResourceMap)).To(Equal(1), "Invalid TS count")
@@ -2303,6 +2321,9 @@ var _ = Describe("Worker Tests", func() {
 				newTS.Spec.VirtualServerAddress = "10.0.0.15"
 				mockCtlr.updateTransportServer(ts, &newTS)
 				mockCtlr.processResources()
+				// Simulating partition priority update to zero by response handler on successfully posting the priority
+				// tenant update
+				mockCtlr.resources.updatePartitionPriority("dev", 0)
 				mockCtlr.processResources()
 				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(2), "Invalid Partition count")
 				Expect(len(mockCtlr.resources.ltmConfig[mockCtlr.Partition].ResourceMap)).To(Equal(2), "Invalid TS count")
@@ -2315,6 +2336,7 @@ var _ = Describe("Worker Tests", func() {
 			var fooEndpts *v1.Endpoints
 			var fooPorts []v1.ServicePort
 			var newEDNS *cisapiv1.ExternalDNS
+			var ts *cisapiv1.TransportServer
 
 			BeforeEach(func() {
 				//Add Virtual Server
@@ -2345,12 +2367,37 @@ var _ = Describe("Worker Tests", func() {
 							},
 						},
 					})
+				ts = test.NewTransportServer(
+					"SampleTS",
+					namespace,
+					cisapiv1.TransportServerSpec{
+						VirtualServerAddress: "10.1.1.1",
+						Pool: cisapiv1.Pool{
+							Service:     "svc1",
+							ServicePort: intstr.IntOrString{StrVal: "port-80"},
+							Monitor: cisapiv1.Monitor{
+								Type:     "tcp",
+								Timeout:  10,
+								Interval: 10,
+							},
+						},
+						BotDefense:         "/Common/bot-defense",
+						DOS:                "/Common/dos",
+						IRules:             []string{"/Common/SampleIRule"},
+						PersistenceProfile: "source-address",
+						AllowVLANs:         []string{"/Common/devtraffic"},
+						Profiles: cisapiv1.ProfileSpec{
+							TCP: cisapiv1.ProfileTCP{
+								Client: "/Common/f5-tcp-lan",
+								Server: "/Common/f5-tcp-wan",
+							},
+							ProfileL4: "/Common/security-fastL4",
+						},
+					},
+				)
 			})
 
 			It("EDNS", func() {
-				//Add Service
-				//go mockCtlr.Agent.agentWorker()
-				//go mockCtlr.Agent.retryWorker()
 				mockCtlr.addEndpoints(fooEndpts)
 				mockCtlr.processResources()
 
@@ -2367,6 +2414,55 @@ var _ = Describe("Worker Tests", func() {
 				Expect(len(mockCtlr.resources.gtmConfig["test"].WideIPs)).To(Equal(0), "EDNS  not deleted")
 
 			})
+
+			It("Process Transport server with EDNS", func() {
+				mockCtlr.addEndpoints(fooEndpts)
+				mockCtlr.processResources()
+
+				svc := test.NewService("svc1", "1", namespace, "NodePort", fooPorts)
+				mockCtlr.addService(svc)
+				mockCtlr.processResources()
+
+				mockCtlr.addEDNS(newEDNS)
+				mockCtlr.processResources()
+				Expect(len(mockCtlr.resources.gtmConfig)).To(Equal(1),
+					"EDNS not processed")
+
+				mockCtlr.TeemData.ResourceType.IPAMTS = make(map[string]int)
+				ts.Spec.Host = "test.com"
+				mockCtlr.addTransportServer(ts)
+				mockCtlr.processResources()
+
+				ts1 := *ts
+
+				ts1.Name = "ts1"
+				ts1.Spec.Host = "test2.com"
+				ts1.Spec.VirtualServerAddress = "10.1.1.2"
+				mockCtlr.addTransportServer(&ts1)
+				mockCtlr.processResources()
+
+				Expect(len(mockCtlr.resources.ltmConfig)).
+					To(Equal(1), "Invalid Partition Count")
+				Expect(len(mockCtlr.resources.ltmConfig["test"].ResourceMap)).
+					To(Equal(2), "Invalid TS Count")
+				Expect(len(mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs["test.com"].Pools[0].Members)).
+					To(Equal(1), "EDNS not processed with Transport Server")
+				Expect(mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs["test.com"].Pools[0].Members[0]).
+					To(Equal("/test/Shared/crd_10_1_1_1_0"),
+						"Invalid EDNS Pool members")
+
+				mockCtlr.deleteTransportServer(ts)
+				mockCtlr.processResources()
+				Expect(len(mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs["test.com"].Pools[0].Members)).
+					To(Equal(0), "Invalid pool member count")
+
+				mockCtlr.deleteEDNS(newEDNS)
+				mockCtlr.processResources()
+				Expect(len(mockCtlr.resources.gtmConfig[DEFAULT_GTM_PARTITION].WideIPs)).
+					To(Equal(0), "EDNS  not deleted")
+
+			})
+
 		})
 
 		Describe("Processing Ingress Link", func() {
@@ -2535,6 +2631,10 @@ var _ = Describe("Worker Tests", func() {
 				ingressLink21.Spec.Partition = "dev"
 				mockCtlr.updateIngressLink(&ingressLink2, &ingressLink21)
 				mockCtlr.processResources()
+				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(2), "Invalid Partition count")
+				// Simulating partition priority update to zero by response handler on successfully posting the priority
+				// tenant update
+				mockCtlr.resources.updatePartitionPriority("dev1", 0)
 				mockCtlr.processResources()
 				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(1), "Invalid Partition count")
 				Expect(len(mockCtlr.resources.ltmConfig["dev"].ResourceMap)).To(Equal(2), "Invalid IL count")
@@ -2552,6 +2652,10 @@ var _ = Describe("Worker Tests", func() {
 				ingressLink11.Spec.Partition = ""
 				mockCtlr.updateIngressLink(ingressLink1, &ingressLink11)
 				mockCtlr.processResources()
+				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(2), "Invalid Partition count")
+				// Simulating partition priority update to zero by response handler on successfully posting the priority
+				// tenant update
+				mockCtlr.resources.updatePartitionPriority("dev", 0)
 				mockCtlr.processResources()
 				Expect(len(mockCtlr.resources.ltmConfig)).To(Equal(1), "Invalid Partition count")
 				Expect(len(mockCtlr.resources.ltmConfig[mockCtlr.Partition].ResourceMap)).To(Equal(2), "Invalid IL count")